% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESB_accnr2id.R
\name{ESB_accnr2id}
\alias{ESB_accnr2id}
\title{Converts accession nr to ESBase id}
\usage{
ESB_accnr2id(accnr)
}
\arguments{
\item{accnr}{An NRM accession nr}
}
\value{
An ESBase id
}
\description{
Converts accession nr to ESBase id
}
\examples{
ESB_accnr2id("A2022/12345");
}
