% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESB_connect.R
\name{ESB_connection}
\alias{ESB_connection}
\title{Connect to ESBase}
\usage{
ESB_connection(user, password)
}
\arguments{
\item{user}{ESBase user name}

\item{password}{ESBase password}
}
\value{
A connection (an rvest_session)
}
\description{
Initiates a connection to ESBase. Note this requires the user to have access to http://esbase.nrm.se, i.e. an account and a connection on the local NRM network.
}
\examples{
\dontrun{
con <- ESB_connection(my_username, my_password)
}
}
