% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESB_deg2dec.R
\name{ESB_deg2dec}
\alias{ESB_deg2dec}
\title{Converts geographical coordinates}
\usage{
ESB_deg2dec(degree)
}
\arguments{
\item{degree}{A text string coordinate}
}
\value{
A decimal coordinate
}
\description{
Converts coordinates of the form "E 19 51,52" to decimal degrees. Mainly for internal use.
}
\examples{
ESB_deg2dec(c("N 63 32,49", "E 19 51,52", "W 19 51,52"))
}
