% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESB_get_selected.R
\name{ESB_get_selected}
\alias{ESB_get_selected}
\title{Get selected item from drop-down menu}
\usage{
ESB_get_selected(node)
}
\arguments{
\item{node}{A html node}
}
\value{
A value
}
\description{
Mainly for internal use
}
