% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESB_projects.R
\name{ESB_projects}
\alias{ESB_projects}
\title{Get projects}
\usage{
ESB_projects(con)
}
\arguments{
\item{con}{A connection initiated by ESB_connect}
}
\value{
A tibble
}
\description{
Downloads a tibble containing details of projects stored in ESBase.
}
\examples{
\dontrun{
con <- ESB_connection(my_username, my_password)
projects <- ESB_projects(con)
}
}
