% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESB_read_accession.R
\name{ESB_read_accession}
\alias{ESB_read_accession}
\title{Gets accession information}
\usage{
ESB_read_accession(con, accnr)
}
\arguments{
\item{con}{A connection initiated by ESB_connect}

\item{accnr}{An accession number}
}
\value{
A tibble
}
\description{
Downloads the accession information of an individual from ESBase
}
\examples{
\dontrun{
con <- ESB_connection(my_username, my_password)
ESB_read_accession(con, "A2022/12345")
}
}
